Fix CLI API import
